V1 Update
Version able to perform simulations.

There is also a Matlab Live Script code that was used to assess the math and prototype the simulations due to the quick visualization capabilities.

Two Gifs from the animation of the simulation performed in matlab, with different initial conditions and spring stiffness.

Yet to add tests.